refactor: RequestLine에 HttpMethod를 물어보도록 개선
