+++ tomcat/src/main/java/org/apache/coyote/http11/request/startLine/HttpMethod.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request.startLine;
 
 import java.util.Arrays;
